add rspec run func into utils file
